Move identifier lookup to event address module
