+++ web/src/lib/List.ts
@@ -3,7 +3,8 @@ import { createRxOneshotReq, latest } from 'rx-nostr';
 import { lastValueFrom } from 'rxjs';
 import type * as Nostr from 'nostr-typedef';
 import { rxNostr, tie } from './timelines/MainTimeline';
-import { filterTags, findIdentifier } from './EventHelper';
+import { filterTags } from './EventHelper';
+import { findIdentifier } from './nostr/protocol/event-address';
 import { isLegacyEncryption } from './nostr/protocol/nip04';
 import { pubkey } from './stores/Author';
 import { Signer } from './Signer';
