+++ web/src/lib/author/PeopleLists.ts
@@ -2,8 +2,7 @@ import { get, writable } from 'svelte/store';
 import { createRxBackwardReq, filterAsync, latestEach, now, uniq } from 'rx-nostr';
 import type * as Nostr from 'nostr-typedef';
 import { isDecodable } from '$lib/Encryption';
-import { findIdentifier } from '$lib/EventHelper';
-import { getEventAddress } from '$lib/nostr/protocol/event-address';
+import { findIdentifier, getEventAddress } from '$lib/nostr/protocol/event-address';
 import { pubkey as authorPubkey } from '$lib/stores/Author';
 import { rxNostr, tie } from '$lib/timelines/MainTimeline';
 import { Signer } from '$lib/Signer';
