+++ web/src/lib/stores/Author.ts
@@ -4,7 +4,8 @@ import type { User } from '../../routes/types';
 import type { Event } from 'nostr-tools';
 import { defaultRelays } from '$lib/Constants';
 import type { Author } from '$lib/Author';
-import { filterTags, findIdentifier } from '$lib/EventHelper';
+import { filterTags } from '$lib/EventHelper';
+import { findIdentifier } from '$lib/nostr/protocol/event-address';
 import { getZapSenderPubkey } from '$lib/nostr/protocol/nip57';
 import { getReadRelays, getWriteRelays, parseRelayList } from '$lib/nostr/protocol/nip65';
 import { decryptListContent } from '$lib/List';
